+++ assets/static/project-map/domains/map-chrome.js
@@ -23,7 +23,7 @@ import {
 } from '../map/map-state.js'
 import { installMapToolbar } from '../map/toolbar.js'
 import { kmToDegreeDeltas } from '../geo.js'
-import { raiseSeekWedgeLayers } from '../map/seek-layers.js'
+import { raiseSeekLensLayers } from '../map/seek-layers.js'
 
 /**
  * Create the MapLibre map, nav control, and toolbar.
@@ -128,7 +128,7 @@ export function createMapChromeDomain(ctx) {
         }
       }
     }
-    raiseSeekWedgeLayers(map)
+    raiseSeekLensLayers(map)
   }
 
   function showTerrainOverlays() {